Fix fail_on_role check
This fixes the issue where the check is never reached because
the module is unconditionally exited, before, if the specified
role doesn't exist.

Which was effectively as if `fail_on_role` is always set to 'no'
plus the silencing of the warning. Which, of course, is quite a trap
and conflicts with the documentation, as well.

For symmetry and better troubleshooting that result now also includes
the executed queries (like the other call site).