Add missing psycopg2.extras import
On some of our systems the `postgresql_query` task failed with the following error:

    AttributeError: module 'psycopg2' has no attribute 'extras'

Applying the patch from this commit fixed the issue for me on those systems. This was previously reported in issue #283. However now that some time has passed, I am unable to reproduce the issue on our newer systems.

In any case, this is still a mistake that could cause problems in the future, so I think it would be better to fix it.

The reason that this mistake is not causing any actual errors in normal situations is because of how the python import system functions: All of the entrypoints (all the source files under plugins/modules) have an import for `psycopg2.extras`. This causes `extras` to be available on `psycopg2` for any code (in any module) that executes after that import statement, such as the function bodies in `module_utils/postgres.py`.

Fixes #283